Rename ExpandedNode files to DocNode for PR comparison
