+++ src/components/DocNode/DocNode.module.scss
